Refactor WeChatClient
