+++ wechatpy/client/api/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, unicode_literals
+
+from .menu import WeChatMenu  # NOQA
+from .user import WeChatUser  # NOQA
+from .card import WeChatCard  # NOQA
+from .group import WeChatGroup  # NOQA
+from .media import WeChatMedia  # NOQA
+from .message import WeChatMessage  # NOQA
+from .qrcode import WeChatQRCode  # NOQA
+from .misc import WeChatMisc  # NOQA
